+++ lib/src/config/smart_config_custom.dart
@@ -52,7 +52,7 @@ class SmartConfigCustom {
   /// centerLeft：dialog位于屏幕左边，动画默认为位移动画，自左而右
   ///
   /// centerRight：dialog位于屏幕左边，动画默认为位移动画，自右而左
-  final AlignmentGeometry alignment;
+  final Alignment alignment;
 
   /// [animationTime]：The animation time can be set
   ///