+++ lib/src/custom/custom_dialog.dart
@@ -24,7 +24,7 @@ class CustomDialog extends BaseDialog {
 
   Future<T?> show<T>({
     required Widget widget,
-    required AlignmentGeometry alignment,
+    required Alignment alignment,
     required bool usePenetrate,
     required bool useAnimation,
     required Duration animationTime,
@@ -98,7 +98,7 @@ class CustomDialog extends BaseDialog {
     required Widget widget,
     required TargetBuilder? targetBuilder,
     required ReplaceBuilder? replaceBuilder,
-    required AlignmentGeometry alignment,
+    required Alignment alignment,
     required bool usePenetrate,
     required bool useAnimation,
     required Duration animationTime,