+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * fix [#142](https://github.com/fluttercandies/flutter_smart_dialog/issues/142)
 * add checkExist
 * fix [#162](https://github.com/fluttercandies/flutter_smart_dialog/issues/162)
+* break change: "AlignmentGeometry" adjust to "Alignment"
 
 
 # [4.9.0]